requirements.txt
bark_ssg==1.3.4
encodec==0.1.1
ipython==8.13.2
nltk==3.8.1
numpy==1.24.3
playsound==1.2.2
PyPDF2==3.0.1
scipy==1.10.1
streamlit==1.22.0
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
tqdm==4.62.3
